Abfrage für das Löschen von News eingefügt

+++ Webpage/blog/views.py
@@ -5,6 +5,8 @@ from django.contrib.auth.decorators import login_required, user_passes_test
 from django.shortcuts import render, redirect, get_object_or_404
 from django.template import Context, Template
 from django.template.base import logger
+
+from utils.member import userToHash
 from .models import blogPost, blogPostHistory
 from django.http import HttpResponse
 from django.contrib.auth.models import Permission
@@ -72,13 +74,23 @@ def adminNewsPage(request):
 @user_passes_test(isUserPartOfGroup_Editor)
 @login_required
 def deleteNews(request):
-    if (request.user.is_authenticated):
-        if ('id' in request.GET):
-            id = request.GET['id']
-            blogPost.objects.get(id=id).delete()
-        return redirect("writerView")
-    else:
-        return redirect("ASV")
+    if ('id' in request.GET) and request.GET['id'] != "" and blogPost.objects.get(id=id).exists():
+
+        id = request.GET['id']
+
+        if 'key' in request.GET and request.GET['key'] != "":
+            # Lösche den Nutzer
+            givenKey = request.GET['key']
+            if givenKey == userToHash(id):
+                blogPost.objects.get(id=id).delete()
+                return redirect("writerView")
+        else:
+            key = userToHash(id)
+            # sende Seite
+            return render(request, "blog/delete.html", {"hash": key, "post": blogPost.objects.get(id=id)})
+
+
+    return redirect("writerView")
 
 '''Insert a new Blog Entry'''
 @user_passes_test(isUserPartOfGroup_Editor)
